+++ NEWS.md
@@ -97,6 +97,7 @@ The changes are reflected in the installation section of the handbook.
 - `annotateSuspects()`: log if the suspect formula/compound data could not be matched with feature annotations
 - Fixed: `chromPeaks()` from `xcms` was somethimes not found (issue #68) 
 - `plotGraph()` method for transformation products gained `width` and `height` arguments.
+- `plotInt()` method for components: `index` argument can now also be component name
 
 
 # patRoon 2.1